Added super constructor to neural net and moved eluc neural net to its own file.
